requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic
boto3
pytest
httpx
# Optional for signature verification on /auth/verify
pynacl
PyJWT
python-dotenv
